ajuste round Report

+++ bingo-api/src/DTOs/Response/report/RoundReportItemDto.cs
@@ -5,14 +5,14 @@ public record RoundReportItemDto
         public DateTime RoundTime { get; set; }
         public int CardSaleCount { get; set; }
         public int BotSaleCount { get; set; }
-        public float Collected { get; set; }
-        public float BotCollected { get; set; }
+        public decimal Collected { get; set; }
+        public decimal BotCollected { get; set; }
         public DateTime? Finished { get; set; }
         public int UserWinners { get; set; }
         public int BotWinners { get; set; }
-        public float UserAwards { get; set; }
-        public float BotAwards { get; set; }
-        public float TotalPrizes { get; set; }
-        public float Comissions { get; set; }
-        public float NetValue { get; set; }
+        public decimal UserAwards { get; set; }
+        public decimal BotAwards { get; set; }
+        public decimal TotalPrizes { get; set; }
+        public decimal Comissions { get; set; }
+        public decimal NetValue { get; set; }
 }
